core: Add send method
This patch adds a single send method in core that can be called like so:

exports.mods.call('ibot_core', 'send', 'JOIN #channel')

Right now this method simply sends the message line to the server passed
in as the second argument, defaulting to the active instance's server if
undefined. This method still needs to fire an event and act accordingly
as detailed in #9.