core: Fire will_send; call own methods using mods
This patch updates the new send method in core to wrap the message to be
sent in an object and then fire an event `will_send`, passing the
wrapped message as an argument. Wrapping the message allows receivers of
the event to update the message as desired.

This patch also updates both Privmsg.reply and core's privmsg method to
make use of exports.mods.call rather than calling in-module methods
directly. This ensures that the correct module is set as the active
module in mods. Strictly speaking, this should not need to be done by
the privsmg method, assuming it itself is called correctly, but it's
better to guard against this just in case.